packages/service/pyproject.toml
[project]
name = "strands-ai-sdk"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "distro==1.9.0",
    "fastapi==0.127.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jiter==0.12.0",
    "openai==2.7.2",
    "psycopg2-binary>=2.9.11",
    "pydantic==2.12.5",
    "sqlmodel>=0.0.29",
    "strands-agents==1.20.0",
    "strands-agents-tools==0.2.18",
    "pydantic_core==2.41.5",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "sniffio==1.3.1",
    "starlette==0.49.0",
    "tqdm==4.67.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "urllib3==2.6.2",
    "uvicorn==0.40.0",
    "vercel==0.3.6",
    "vercel-sandbox==0.0.3",
    "vercel-sdk==0.0.8",
    "python-jose[cryptography]==3.5.0",
    "oic==1.7.0",
    "cachetools>=6.2.4",
]

[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
